pyproject.toml
[project]
name = "loan-data-service"
version = "0.1.0"
description = "Loan data ingestion service + MCP agent surface"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "python-multipart>=0.0.9",
    "sqlparse>=0.5",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "mcp>=1.2",
    "anthropic>=0.40",
    "google-genai>=0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loan_api", "src/loan_mcp"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
